% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocation_summary.R
\name{allocation_summary}
\alias{allocation_summary}
\title{Allocation Summary}
\usage{
allocation_summary(schedule, group_col = "Group")
}
\arguments{
\item{schedule}{A data frame or tibble produced by an
ExpDesignR randomization function.}

\item{group_col}{Name of the treatment group column.}
}
\value{
A tibble summarizing the number and percentage
of subjects in each treatment group.
}
\description{
Summarizes treatment allocations from a randomization schedule.
}
\examples{
sch <- simple_randomization(
  n = 20,
  groups = c("Control","Treatment"),
  seed = 123
)

allocation_summary(sch)

}
