% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_additional.R
\name{completely_randomized_design}
\alias{completely_randomized_design}
\title{Completely Randomized Design}
\usage{
completely_randomized_design(n, treatments, seed = NULL, ratio = NULL)
}
\arguments{
\item{n}{Number of units.}

\item{treatments}{Treatment labels.}

\item{seed}{Optional random seed.}

\item{ratio}{Optional allocation weights.}
}
\value{
A tibble containing unit and treatment.
}
\description{
Randomly assign experimental units to treatment groups.
}
