% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latin_square.R
\name{latin_square}
\alias{latin_square}
\title{Latin Square Design}
\usage{
latin_square(treatments, randomize = TRUE, seed = NULL)
}
\arguments{
\item{treatments}{Character vector of treatment labels.}

\item{randomize}{Logical. Should rows, columns and treatments
be randomized? Default is TRUE.}

\item{seed}{Optional random seed.}
}
\value{
A matrix representing a Latin square.
}
\description{
Generates a Latin Square design for experimental studies.
}
\examples{
latin_square(
  treatments = LETTERS[1:4],
  seed = 123
)

}
