% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_randomization.R
\name{plot_randomization}
\alias{plot_randomization}
\title{Plot Randomization Schedule}
\usage{
plot_randomization(
  schedule,
  group_col = "Group",
  fill = "#2C7FB8",
  title = "Treatment Allocation"
)
}
\arguments{
\item{schedule}{A data frame produced by ExpDesignR.}

\item{group_col}{Character. Name of the treatment column.}

\item{fill}{Character. Fill colour.}

\item{title}{Character. Plot title.}
}
\value{
A ggplot object.
}
\description{
Creates a bar chart showing the number of subjects allocated
to each treatment group.
}
\examples{
sch <- simple_randomization(
  n = 40,
  groups = c("Control","Treatment"),
  seed = 123
)

plot_randomization(sch)

}
