% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designs_additional.R
\name{split_plot_design}
\alias{split_plot_design}
\title{Split-Plot Design}
\usage{
split_plot_design(whole, sub, n_whole = length(whole), seed = NULL)
}
\arguments{
\item{whole}{Whole-plot treatment labels.}

\item{sub}{Sub-plot treatment labels.}

\item{n_whole}{Number of whole plots.}

\item{seed}{Optional random seed.}
}
\value{
A tibble with whole plots and sub-plots.
}
\description{
Generate a randomized split-plot treatment schedule.
}
