% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_schedule.R
\name{export_schedule}
\alias{export_schedule}
\title{Export Randomization Schedule}
\usage{
export_schedule(schedule, file, row.names = FALSE)
}
\arguments{
\item{schedule}{A data frame or tibble generated by ExpDesignR.}

\item{file}{Character. Output CSV filename or path. This argument
must be supplied explicitly.}

\item{row.names}{Logical. Should row names be written?}
}
\value{
Invisibly returns the input \code{schedule} unchanged.
  The function writes the schedule to the CSV file specified by
  \code{file}.
}
\description{
Export a randomization schedule to a CSV file.
}
\examples{
sch <- simple_randomization(
  n = 20,
  groups = c("Control", "Treatment"),
  seed = 123
)

tf <- tempfile(fileext = ".csv")

export_schedule(
  sch,
  file = tf
)

unlink(tf)

}
